+++ weaviate/collection/classes/config.py
@@ -396,98 +396,6 @@ class _CollectionConfig:
     vectorizer: Vectorizer
 
 
-def _collection_config_from_json(schema: Dict[str, Any]) -> _CollectionConfig:
-    def _is_primitive(d_type: str) -> bool:
-        return d_type[0][0].lower() == d_type[0][0]
-
-    def _property_data_type_from_weaviate_data_type(
-        data_type: List[str],
-    ) -> Union[DataType, _ReferenceDataType, _ReferenceDataTypeMultiTarget]:
-        if len(data_type) == 1 and _is_primitive(data_type[0]):
-            return DataType(data_type[0])
-
-        if len(data_type) == 1:
-            return _ReferenceDataType(target_collection=data_type[0])
-
-        return _ReferenceDataTypeMultiTarget(target_collections=data_type)
-
-    return _CollectionConfig(
-        name=schema["class"],
-        description=schema.get("description"),
-        inverted_index_config=_InvertedIndexConfig(
-            bm25=_BM25Config(
-                b=schema["invertedIndexConfig"]["bm25"]["b"],
-                k1=schema["invertedIndexConfig"]["bm25"]["k1"],
-            ),
-            cleanup_interval_seconds=schema["invertedIndexConfig"]["cleanupIntervalSeconds"],
-            stopwords=_StopwordsConfig(
-                preset=StopwordsPreset(schema["invertedIndexConfig"]["stopwords"]["preset"]),
-                additions=schema["invertedIndexConfig"]["stopwords"]["additions"],
-                removals=schema["invertedIndexConfig"]["stopwords"]["removals"],
-            ),
-        ),
-        multi_tenancy_config=_MultiTenancyConfig(enabled=schema["multiTenancyConfig"]["enabled"]),
-        properties=[
-            _Property(
-                data_type=_property_data_type_from_weaviate_data_type(prop["dataType"]),
-                description=prop.get("description"),
-                index_filterable=prop["indexFilterable"],
-                index_searchable=prop["indexSearchable"],
-                name=prop["name"],
-                tokenization=Tokenization(prop["tokenization"])
-                if prop.get("tokenization") is not None
-                else None,
-            )
-            for prop in schema["properties"]
-        ]
-        if schema.get("properties") is not None
-        else [],
-        replication_factor=_ReplicationFactor(factor=schema["replicationConfig"]["factor"]),
-        sharding_config=_ShardingConfig(
-            virtual_per_physical=schema["shardingConfig"]["virtualPerPhysical"],
-            desired_count=schema["shardingConfig"]["desiredCount"],
-            actual_count=schema["shardingConfig"]["actualCount"],
-            desired_virtual_count=schema["shardingConfig"]["desiredVirtualCount"],
-            actual_virtual_count=schema["shardingConfig"]["actualVirtualCount"],
-            key=schema["shardingConfig"]["key"],
-            strategy=schema["shardingConfig"]["strategy"],
-            function=schema["shardingConfig"]["function"],
-        ),
-        vector_index_config=_VectorIndexConfig(
-            cleanup_interval_seconds=schema["vectorIndexConfig"]["cleanupIntervalSeconds"],
-            distance=VectorDistance(schema["vectorIndexConfig"]["distance"]),
-            dynamic_ef_min=schema["vectorIndexConfig"]["dynamicEfMin"],
-            dynamic_ef_max=schema["vectorIndexConfig"]["dynamicEfMax"],
-            dynamic_ef_factor=schema["vectorIndexConfig"]["dynamicEfFactor"],
-            ef=schema["vectorIndexConfig"]["ef"],
-            ef_construction=schema["vectorIndexConfig"]["efConstruction"],
-            flat_search_cutoff=schema["vectorIndexConfig"]["flatSearchCutoff"],
-            max_connections=schema["vectorIndexConfig"]["maxConnections"],
-            pq=_PQConfig(
-                enabled=schema["vectorIndexConfig"]["pq"]["enabled"],
-                bit_compression=schema["vectorIndexConfig"]["pq"]["bitCompression"],
-                segments=schema["vectorIndexConfig"]["pq"]["segments"],
-                centroids=schema["vectorIndexConfig"]["pq"]["centroids"],
-                training_limit=schema["vectorIndexConfig"]["pq"]["trainingLimit"],
-                encoder=_PQEncoderConfig(
-                    type_=PQEncoderType(schema["vectorIndexConfig"]["pq"]["encoder"]["type"]),
-                    distribution=PQEncoderDistribution(
-                        schema["vectorIndexConfig"]["pq"]["encoder"]["distribution"]
-                    ),
-                ),
-            ),
-            skip=schema["vectorIndexConfig"]["skip"],
-            vector_cache_max_objects=schema["vectorIndexConfig"]["vectorCacheMaxObjects"],
-        ),
-        vector_index_type=VectorIndexType(schema["vectorIndexType"]),
-        vectorizer=Vectorizer(schema["vectorizer"]),
-    )
-
-
-def _collection_configs_from_json(schema: Dict[str, Any]) -> Dict[str, _CollectionConfig]:
-    return {schema["class"]: _collection_config_from_json(schema) for schema in schema["classes"]}
-
-
 # class PropertyConfig(ConfigCreateModel):
 #     indexFilterable: Optional[bool] = Field(None, alias="index_filterable")
 #     indexSearchable: Optional[bool] = Field(None, alias="index_searchable")
@@ -529,22 +437,6 @@ class Property(ConfigCreateModel):
         ret_dict = super().to_dict()
         ret_dict["dataType"] = [ret_dict["dataType"]]
         return ret_dict
-
-
-class ReferenceDataType(BaseModel):
-    """Use this class when defining the collections to which your collection should cross reference when specifying
-    a reference property.
-
-    A cross reference can refer to multiple collections at once, hence the ability to specify a list of collections.
-
-    However, be careful when doing so as all downstream inserts, queries, and searches for this collection must specify
-    the exact collection to which the reference refers.
-
-    If you are unsure, it is recommended to use a single collection
-    for each cross reference.
-    """
-
-    collections: Union[str, List[str]]
 
 
 class ReferencePropertyBase(ConfigCreateModel):
