+++ weaviate/collection/collection_base.py
@@ -4,9 +4,9 @@ from requests.exceptions import ConnectionError as RequestsConnectionError
 
 from weaviate.collection.classes.config import (
     CollectionConfigCreateBase,
-    _collection_configs_from_json,
     _CollectionConfig,
 )
+from weaviate.collection.classes.config_methods import _collection_configs_from_json
 from weaviate.connect import Connection
 from weaviate.exceptions import UnexpectedStatusCodeException
