+++ weaviate/collection/data.py
@@ -10,8 +10,8 @@ from weaviate.collection.classes.data import (
     DataObject,
     Error,
     ReferenceTo,
-    GetObjectByIdIncludes,
-    GetObjectsIncludes,
+    GetObjectByIdMetadata,
+    GetObjectsMetadata,
     IncludesModel,
     ReferenceToMultiTarget,
     _BatchReturn,
@@ -145,7 +145,7 @@ class _Data:
         raise UnexpectedStatusCodeException("Update object", response)
 
     def _get_by_id(
-        self, uuid: UUID, includes: Optional[GetObjectByIdIncludes] = None
+        self, uuid: UUID, includes: Optional[GetObjectByIdMetadata] = None
     ) -> Optional[Dict[str, Any]]:
         path = f"/objects/{self.name}/{uuid}"
 
@@ -153,7 +153,7 @@ class _Data:
             params=self.__apply_context({}), path=path, includes=includes
         )
 
-    def _get(self, includes: Optional[GetObjectsIncludes] = None) -> Optional[Dict[str, Any]]:
+    def _get(self, includes: Optional[GetObjectsMetadata] = None) -> Optional[Dict[str, Any]]:
         path = "/objects"
         params: Dict[str, Any] = {"class": self.name}
 
@@ -354,14 +354,14 @@ class _DataCollection(_Data):
         self._update(weaviate_obj, uuid=uuid)
 
     def get_by_id(
-        self, uuid: UUID, includes: Optional[GetObjectByIdIncludes] = None
+        self, uuid: UUID, includes: Optional[GetObjectByIdMetadata] = None
     ) -> Optional[_Object]:
         ret = self._get_by_id(uuid=uuid, includes=includes)
         if ret is None:
             return ret
         return self._json_to_object(ret)
 
-    def get(self, includes: Optional[GetObjectsIncludes] = None) -> List[_Object]:
+    def get(self, includes: Optional[GetObjectsMetadata] = None) -> List[_Object]:
         ret = self._get(includes=includes)
         if ret is None:
             return []
@@ -490,15 +490,15 @@ class _DataCollectionModel(Generic[Model], _Data):
         self._update(weaviate_obj, uuid)
 
     def get_by_id(
-        self, uuid: UUID, includes: Optional[GetObjectByIdIncludes] = None
+        self, uuid: UUID, metadata: Optional[GetObjectByIdMetadata] = None
     ) -> Optional[_Object[Model]]:
-        ret = self._get_by_id(uuid=uuid, includes=includes)
+        ret = self._get_by_id(uuid=uuid, includes=metadata)
         if ret is None:
             return None
         return self._json_to_object(ret)
 
-    def get(self, includes: Optional[GetObjectsIncludes] = None) -> List[_Object[Model]]:
-        ret = self._get(includes=includes)
+    def get(self, metadata: Optional[GetObjectsMetadata] = None) -> List[_Object[Model]]:
+        ret = self._get(includes=metadata)
         if ret is None:
             return []
