+++ weaviate/collection/grpc.py
@@ -1,5 +1,7 @@
 import datetime
+from dataclasses import dataclass
 from typing import Optional, List, Dict, Union, Set, Protocol, Generic, cast, Tuple
+from typing_extensions import TypeAlias
 
 import grpc
 import uuid as uuid_lib
@@ -23,13 +25,8 @@ from weaviate.collection.classes.grpc import (
     PROPERTIES,
     LinkTo,
     LinkToMultiTarget,
-    GrpcResult,
-    SearchResponse,
     RefProps,
-    _StructValue,
     ReturnValues,
-    _PyValue,
-    _RawObject,
 )
 from weaviate.collection.classes.internal import (
     _MetadataReturn,
@@ -41,6 +38,52 @@ from weaviate.connect import Connection
 from weaviate.exceptions import WeaviateGRPCException
 from weaviate.weaviate_types import UUID
 from weaviate_grpc import weaviate_pb2
+
+# Can be found in the google.protobuf.internal.well_known_types.pyi stub file but is defined explicitly here for clarity.
+_StructValue: TypeAlias = Union[
+    struct_pb2.Struct,
+    struct_pb2.ListValue,
+    str,
+    float,
+    bool,
+    None,
+    List[float],
+    List[int],
+    List[str],
+    List[bool],
+    List[UUID],
+]
+_PyValue: TypeAlias = Union[
+    Dict[str, "_PyValue"],
+    List["_PyValue"],
+    str,
+    float,
+    bool,
+    None,
+    List[float],
+    List[int],
+    List[str],
+    List[bool],
+    List[UUID],
+]
+_RawObject = Dict[str, _PyValue]
+
+
+@dataclass
+class GrpcResult:
+    metadata: _MetadataReturn
+    result: Dict[str, Union[_StructValue, List["GrpcResult"]]]
+
+
+@dataclass
+class SearchResult:
+    properties: weaviate_pb2.ResultProperties
+    additional_properties: weaviate_pb2.ResultAdditionalProps
+
+
+@dataclass
+class SearchResponse:
+    results: List[SearchResult]
 
 
 class _GRPC:
